Ensure environment variables are set to run the scenario
